% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{gist.csv}
\alias{gist.csv}
\title{Creates new gist on GitHub using the Authentication token found in
~/.gist-vim file. This is compliant with the Vim Gist plugin so if you're
already using that one, you'll be set to use the R package too. By default
gists created here are private, but this and description could be set. If you
give a gist id as argument to the function, it will try to update an existing
gist instead of creating a new one.}
\usage{
gist.csv(DT, public = FALSE, description = "", id = NULL)
}
\description{
Creates new gist on GitHub using the Authentication token found in
~/.gist-vim file. This is compliant with the Vim Gist plugin so if you're
already using that one, you'll be set to use the R package too. By default
gists created here are private, but this and description could be set. If you
give a gist id as argument to the function, it will try to update an existing
gist instead of creating a new one.
}

