% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{gist.csv}
\alias{gist.csv}
\title{Creates new gist on GitHub using the Authentication token found in
~/.gist-vim file. This is compliant with the Vim Gist plugin so if you're
already using that one, you'll be set to use the R package too. By default
gists created here are private, but this and description could be set. If you
give a gist id as argument to the function, it will try to update an existing
gist instead of creating a new one.}
\usage{
gist.csv(DT, public = FALSE, description = "", id = NULL,
  csv.writer = write.csv, row.names = FALSE, ...)
}
\arguments{
\item{public}{should the gist be made public or secrete on http://gist.github.com}

\item{description}{the description for the new gist.}

\item{id}{if left NULL a new gist will be created. If given a gist ID string,
an existing gist will be updated.}

\item{csv.writer}{this takes a function as a value e.g. write.csv or write.csv2. See README.md.}

\item{...}{additional params passed on to csv.writer function.}

\item{raw.names}{if this is set to TRUE, then raw.names will be written to
the output.}
}
\description{
Creates new gist on GitHub using the Authentication token found in
~/.gist-vim file. This is compliant with the Vim Gist plugin so if you're
already using that one, you'll be set to use the R package too. By default
gists created here are private, but this and description could be set. If you
give a gist id as argument to the function, it will try to update an existing
gist instead of creating a new one.
}
